requirements.txt
aioredis==2.0.1
annotated-types==0.7.0
anyio==4.6.2.post1
async-timeout==5.0.1
asyncpg==0.30.0
click==8.1.7
contourpy==1.3.1
cycler==0.12.1
databases==0.9.0
fastapi==0.115.5
fonttools==4.55.0
greenlet==3.1.1
h11==0.14.0
idna==3.10
Jinja2==3.1.4
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.9.2
mysql-connector-python==9.1.0
numpy==2.1.3
packaging==24.2
pillow==11.0.0
psycopg2-binary==2.9.10
pydantic==2.10.1
pydantic_core==2.27.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-multipart==0.0.17
redis==5.2.0
setuptools==75.6.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
typing_extensions==4.12.2
uvicorn==0.32.1
